packages/python/contextcompany/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contextcompany"
version = "0.1.0"
description = "The Context Company AI Agent Observability SDK for Python"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "The Context Company", email = "support@thecontext.company"},
]
keywords = ["observability", "ai", "agents", "langchain", "llm", "tracing", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "requests>=2.31.0",
]

[project.optional-dependencies]
langchain = [
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-langchain>=0.47.3",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://thecontext.company"
Documentation = "https://docs.thecontext.company"
Repository = "https://github.com/The-Context-Company/observatory"
Issues = "https://github.com/The-Context-Company/observatory/issues"

[tool.hatch.build.targets.wheel]
packages = ["contextcompany"]
